% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aufgaben.R
\name{Aufgaben}
\alias{Aufgaben}
\alias{Aufgaben_2}
\alias{Aufgaben_7}
\title{Aufgaben vorbereiten

Mit diesen Funktioenn führen Sie die notwendigen Schritte durch, um die Aufgabenbearbeitung vorzubereiten.}
\usage{
Aufgaben_2()

Aufgaben_7()
}
\description{
Aufgaben vorbereiten

Mit diesen Funktioenn führen Sie die notwendigen Schritte durch, um die Aufgabenbearbeitung vorzubereiten.
}
