% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitzungen.R
\name{Sitzung_2}
\alias{Sitzung_2}
\title{Übungen für das Praktikum in PsyBSc 7, Sommersemester 2020}
\usage{
Sitzung_2()
}
\description{
Mit diesen Funktionen können Sie die Übungen zu den einzelnen Sitzungen aufrufen.
}
