% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitzungen.R
\name{Sitzungen}
\alias{Sitzungen}
\alias{Sitzung_1}
\alias{Sitzung_2}
\alias{Sitzung_3}
\title{Übungen für das Praktikum in PsyBSc 7, Sommersemester 2020

Mit diesen Funktionen können Sie die Übungen zu den einzelnen Sitzungen aufrufen. Die Funktionen nehmen keine Argumente entgegen.}
\usage{
Sitzung_1()

Sitzung_2()

Sitzung_3()
}
\description{
Übungen für das Praktikum in PsyBSc 7, Sommersemester 2020

Mit diesen Funktionen können Sie die Übungen zu den einzelnen Sitzungen aufrufen. Die Funktionen nehmen keine Argumente entgegen.
}
